requirements.txt
# Core ML
torch>=2.2
transformers>=4.40
peft>=0.10
datasets>=2.19
accelerate>=0.29
bitsandbytes>=0.43

# Evaluation
rouge-score>=0.1.2
numpy>=1.26

# API
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.7

# Configuration and utilities
python-dotenv>=1.0
pyyaml>=6.0
tqdm>=4.66

# Development / testing
pytest>=8.2
httpx>=0.27
